+++ functions/package.json
@@ -14,6 +14,7 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "express": "^4.17.1",
     "firebase-admin": "^9.4.2",
     "firebase-functions": "^3.11.0",
     "firebase-tools": "^9.1.0"
