+++ functions/src/index.ts
@@ -1,238 +1,43 @@
 import * as functions from "firebase-functions";
 import * as admin from "firebase-admin";
-import { SingleTargetRequest, SingleTargetResult } from "./models/SingleTarget";
-import { gameConverter } from "./models/Game";
-import { Player } from "./models/Player";
-import { KickPlayer } from "./models/HostEvents";
 
 admin.initializeApp();
 
-import FirestoreUtil from "./helper/FirestoreUtil";
+import { kickPlayerHandler } from "./https/KickPlayer";
+import { closeGameHandler } from "./https/CloseGame";
+import { garbageCollectionHTTPSHandler } from "./https/GarbageCollection";
+import { garbageCollectionHandler } from "./timed/GarbageCollection";
+import { onPlayerLeaveHandler } from "./firestore/OnPlayerLeave";
+import { onPlayerJoinHandler } from "./firestore/OnPlayerJoin";
+import { singleTargetHandler } from "./https/SingleTarget";
 
-const db = admin.firestore();
-
-export const singleTarget = functions.region("europe-west1").https.onCall(
-  async (data: SingleTargetRequest, context): Promise<SingleTargetResult> => {
-    const auth = context.auth;
-    if (auth) {
-      const requestUID = auth.uid;
-
-      const gameData = await FirestoreUtil.getGameData(data.gameID);
-      if (!gameData) {
-        throw new functions.https.HttpsError(
-          "not-found",
-          "Game data was missing!"
-        );
-      }
-      if (requestUID !== gameData.taskTarget) {
-        throw new functions.https.HttpsError(
-          "permission-denied",
-          "Function must be called by target of the question!"
-        );
-      }
-
-      const playerData = await FirestoreUtil.getPlayerData(
-        data.gameID,
-        requestUID
-      );
-      if (!playerData) {
-        throw new functions.https.HttpsError(
-          "not-found",
-          "Player data was missing!"
-        );
-      }
-
-      await FirestoreUtil.getGame(data.gameID).update({
-        evalState: true,
-      });
-
-      await FirestoreUtil.getPlayer(data.gameID, requestUID).set(
-        new Player(
-          playerData.uid,
-          playerData.nickname,
-          data.answer ? playerData.sips : playerData.sips + gameData.penalty,
-          null
-        )
-      );
-
-      return { status: "All okay", responseOK: true };
-    } else {
-      throw new functions.https.HttpsError(
-        "unauthenticated",
-        "User was not authenticated with firebase while calling this function!"
-      );
-    }
-  }
-);
+export const singleTarget = functions
+  .region("europe-west1")
+  .https.onCall(singleTargetHandler);
 
 export const onPlayerJoin = functions
   .region("europe-west1")
   .firestore.document("/games/{gameID}/players/{playerID}")
-  .onCreate(async (snapshot, context) => {
-    if (context.params.playerID === "register") {
-      // If the created document was a register don't proceed!
-      throw new functions.https.HttpsError(
-        "ok",
-        "Function triggered on register create"
-      );
-    }
-
-    const registerRef = await FirestoreUtil.getRegisterRef(
-      context.params.gameID
-    );
-    const playerData = snapshot.data();
-
-    let playerUid: Map<string, string> = new Map();
-
-    if (registerRef.exists) {
-      playerUid = FirestoreUtil.createMap(registerRef);
-    }
-
-    playerUid.set(context.params.playerID, playerData.nickname);
-
-    await FirestoreUtil.updateRegister(context.params.gameID, playerUid);
-
-    return null;
-  });
+  .onCreate(onPlayerJoinHandler);
 
 export const onPlayerLeave = functions
   .region("europe-west1")
   .firestore.document("/games/{gameID}/players/{playerID}")
-  .onDelete(async (snapshot, context) => {
-    if (context.authType === "ADMIN") {
-      // If the delete was performed by an ADMIN (console or firebase) ignore it
-      throw new functions.https.HttpsError("ok", "Delete by admin is ignored!");
-    }
-
-    const registerRef = await FirestoreUtil.getRegisterRef(
-      context.params.gameID
-    );
-
-    const map = FirestoreUtil.createMap(registerRef);
-    map.delete(context.params.playerID);
-
-    await FirestoreUtil.updateRegister(context.params.gameID, map);
-
-    return null;
-  });
+  .onDelete(onPlayerLeaveHandler);
 
 export const garbageCollection = functions
   .region("europe-west1")
   .pubsub.schedule("every 12 hours")
-  .onRun(async () => {
-    const maxAge = Date.now() - 12 * 60 * 60 * 1000;
-    const gamesRef = await db
-      .collection("games")
-      .where("created", "<", new Date(maxAge))
-      .withConverter(gameConverter)
-      .get();
-    await gamesRef.forEach(async (gameToDelete) => {
-      const players = await FirestoreUtil.getPlayers(gameToDelete.id).get();
-      await players.forEach(async (playerToDelete) => {
-        console.log("Player, ", playerToDelete.id);
-        await playerToDelete.ref.delete();
-      });
-
-      await gameToDelete.ref.delete();
-      await db.collection("tictactoe").doc(gameToDelete.id).delete();
-    });
-  });
+  .onRun(garbageCollectionHandler);
 
 export const garbageCollectionHTTPS = functions
   .region("europe-west1")
-  .https.onRequest(async (req, resp) => {
-    const maxAge = Date.now() - 12 * 60 * 60 * 1000;
-    const gamesRef = await db
-      .collection("games")
-      .where("created", "<", new Date(maxAge))
-      .withConverter(gameConverter)
-      .get();
-    const results: string[] = [];
-    await gamesRef.forEach(async (gameToDelete) => {
-      results.push(gameToDelete.id);
-
-      const players = await FirestoreUtil.getPlayers(gameToDelete.id).get();
-      await players.forEach(async (playerToDelete) => {
-        console.log("Player, ", playerToDelete.id);
-        await playerToDelete.ref.delete();
-      });
-
-      await gameToDelete.ref.delete();
-    });
-    resp.json(results);
-  });
+  .https.onRequest(garbageCollectionHTTPSHandler);
 
 export const closeGame = functions
   .region("europe-west1")
-  .https.onCall(async (data: { gameID: string }, context) => {
-    const auth = context.auth;
-    if (auth) {
-      const requestUID = auth.uid;
-      const gameData = await FirestoreUtil.getGameData(data.gameID);
-
-      if (!gameData) {
-        throw new functions.https.HttpsError(
-          "not-found",
-          "Game data was missing!"
-        );
-      }
-      if (requestUID !== gameData.host) {
-        throw new functions.https.HttpsError(
-          "permission-denied",
-          "Function must be called by host of game!"
-        );
-      }
-      const gameRef = FirestoreUtil.getGame(data.gameID);
-
-      const players = await FirestoreUtil.getPlayers(data.gameID).get();
-      await players.forEach(async (playerToDelete) => {
-        console.log("Player, ", playerToDelete.id);
-        await playerToDelete.ref.delete();
-      });
-
-      await gameRef.delete();
-      await db.collection("tictactoe").doc(data.gameID).delete();
-    } else {
-      throw new functions.https.HttpsError(
-        "unauthenticated",
-        "User was not authenticated with firebase while calling this function!"
-      );
-    }
-  });
+  .https.onCall(closeGameHandler);
 
 export const kickPlayer = functions
   .region("europe-west1")
-  .https.onCall(async (data: KickPlayer, context) => {
-    const auth = context.auth;
-    if (auth) {
-      const requestUID = auth.uid;
-      const gameData = await FirestoreUtil.getGameData(data.gameID);
-
-      if (!gameData) {
-        throw new functions.https.HttpsError(
-          "not-found",
-          "Game data was missing!"
-        );
-      }
-      if (requestUID !== gameData.host) {
-        throw new functions.https.HttpsError(
-          "permission-denied",
-          "Function must be called by host of game!"
-        );
-      }
-
-      await FirestoreUtil.getPlayer(data.gameID, data.playerID).delete();
-
-      const registerRef = await FirestoreUtil.getRegisterRef(data.gameID);
-      const map = FirestoreUtil.createMap(registerRef);
-
-      map.delete(data.playerID);
-
-      await FirestoreUtil.updateRegister(data.gameID, map);
-    } else {
-      throw new functions.https.HttpsError(
-        "unauthenticated",
-        "User was not authenticated with firebase while calling this function!"
-      );
-    }
-  });
+  .https.onCall(kickPlayerHandler);
